// src/enums.ts
export enum ESauce {
  tomato = 'tomato',
  alfredo = 'alfredo',
}

export enum ECheese {
  mozzarella = 'mozzarella',
  cheddar = 'cheddar',
  parmesan = 'parmesan',
}

export enum EMeat {
  pepperoni = 'pepperoni',
  ham = 'ham',
  bacon = 'bacon',
  chicken = 'chicken',
  shrimp = 'shrimp',
}

export enum EVeggie {
  tomatoes = 'tomatoes',
  onions = 'onions',
  bell_peppers = 'bell_peppers',
  mushrooms = 'mushrooms',
  black_olives = 'black_olives',
  pineapples = 'pineapples',
}